move toNumberOrZero to numbers.js, so we can remove numberUtils.js
