--- misc/services/mobile-webui/mobile-webui-frontend/src/utils/numberUtils.js
@@ -1,7 +1,0 @@
-export const toNumberOrZero = (arg) => {
-  try {
-    return Number(arg);
-  } catch (e) {
-    return 0;
-  }
-};
