+++ misc/services/mobile-webui/mobile-webui-frontend/src/utils/numbers.js
@@ -1,3 +1,11 @@
 export const round = (number, precision) => {
   return parseFloat(Number(number).toFixed(precision));
 };
+
+export const toNumberOrZero = (arg) => {
+  try {
+    return Number(arg);
+  } catch (e) {
+    return 0;
+  }
+};
